git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@131 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ plugins/uisystem/Event.h
@@ -324,7 +324,7 @@ namespace hare_ui
             return eventType == uiEVT_LEAVE_WINDOW;
         }
 
-        int getWheelDelta() const 
+        f32 getWheelDelta() const 
         {
             return wheelDelta; 
         }
@@ -353,7 +353,7 @@ namespace hare_ui
         bool altDown;
         bool metaDown;
 
-        int wheelDelta;
+        f32 wheelDelta;
         PointF position;
     };
